refactor: locales/HOC to styleguide and out of pkg
