+++ package/src/utils/index.js
@@ -5,4 +5,3 @@ export { default as preventAccidentalDoubleClick } from "./preventAccidentalDoub
 export { default as getRequiredValidator } from "./getRequiredValidator";
 export { default as getPhoneNumberValidator } from "./getPhoneNumberValidator";
 export { default as withStripeElements } from "./withStripeElements";
-export { default as withDefaultLocales } from "./withDefaultLocales";
