+++ styleguide/src/components/utils/locales/locales.json
